requirements.txt
certifi==2019.11.28
chardet==3.0.4
defusedxml==0.6.0
Django==2.2.7
django-allauth==0.41.0
django-ckeditor==5.8.0
django-classy-tags==0.9.0
django-cms==3.7.1
django-crispy-forms==1.8.1
django-formtools==2.2
django-js-asset==1.2.2
django-sekizai==1.0.0
django-treebeard==4.3.1
djangocms-admin-style==1.4.0
djangocms-text-ckeditor==3.8.0
gunicorn==20.0.4
html5lib==1.0.1
idna==2.8
oauthlib==3.1.0
Pillow==7.0.0
psycopg2-binary==2.8.4
PyJWT==1.7.1
python3-openid==3.1.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
six==1.13.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
sorl-thumbnail==12.5.0
sqlparse==0.3.0
urllib3==1.25.7
webencodings==0.5.1
whitenoise==5.0.1
